chore: format files

+++ src/swap/Btc_Marketplace.sol
@@ -157,7 +157,8 @@ contract BtcMarketPlace {
         relay.validateProof(transaction, proof);
 
         // Check output script pubkey (recipient address) and amount
-        uint txOutputValue = BitcoinTx.getTxOutputValue(keccak256(accept.bitcoinAddress.scriptPubKey), transaction.outputVector);
+        uint256 txOutputValue =
+            BitcoinTx.getTxOutputValue(keccak256(accept.bitcoinAddress.scriptPubKey), transaction.outputVector);
         assert(txOutputValue >= accept.amountBtc);
 
         IERC20(accept.ercToken).safeTransfer(accept.requester, accept.ercAmount);
@@ -256,9 +257,9 @@ contract BtcMarketPlace {
 
         BtcBuyOrder storage order = btcBuyOrders[accept.orderId];
         // Check output script pubkey (recipient address) and amount
-        uint txOutputValue = BitcoinTx.getTxOutputValue(keccak256(order.bitcoinAddress.scriptPubKey), transaction.outputVector);
+        uint256 txOutputValue =
+            BitcoinTx.getTxOutputValue(keccak256(order.bitcoinAddress.scriptPubKey), transaction.outputVector);
         assert(txOutputValue >= order.amountBtc);
-
 
         IERC20(accept.ercToken).safeTransfer(accept.accepter, accept.ercAmount);
